+++ tests/test_product.py
@@ -1,3 +1,5 @@
+import datetime
+
 import pytest
 
 from saleor.product import models
@@ -16,3 +18,13 @@ def test_stock_allocator(product_in_stock):
     models.Stock.objects.allocate_stock(stock, 1)
     stock = models.Stock.objects.get(pk=stock.pk)
     assert stock.quantity_allocated == 1
+
+
+def test_product_preview(admin_client, client, product_in_stock):
+    product_in_stock.available_on = (
+        datetime.date.today() + datetime.timedelta(days=7))
+    product_in_stock.save()
+    response = client.get(product_in_stock.get_absolute_url())
+    assert response.status_code == 404
+    response = admin_client.get(product_in_stock.get_absolute_url())
+    assert response.status_code == 200
